+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `ActiveNode` models now respond to `update_all` (thanks ProGM / #1113)
 - Association chains now respond to `update_all` and `update_all_rels` (thanks ProGM / #1113)
 - Rails will now rescue all `Neo4j::RecordNotFound` errors with a 404 status code by default
+- A clone of [ActiveRecord::Enum](http://edgeapi.rubyonrails.org/classes/ActiveRecord/Enum.html) API. See docs for details. (thanks ProGM / #1129)
 
 ### Removed